# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Construction Enterprise Operating Platform API
  version: "0.8.1"
  description: Unified enterprise operating platform for construction company governance, business portal, field OS, and AI governance.
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - 
    url: "http://localhost:3000"
    description: Local development
  - 
    url: "http://localhost:8080"
    description: Docker compose
  - 
    url: "https://ceop.mirai-dx-platform.com"
    description: Production
...
paths:
...
  /api/v1/roles:
    get:
      operationId: listRoles
      summary: Paginated list of roles
      tags:
        - Roles
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/limitParam"
        - 
          $ref: "#/components/parameters/offsetParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - roles
                  - count
                  - total
                  - limit
                  - offset
                properties:
                  roles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"
                  count:
                    type: integer
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createRole
      summary: Create a role
      tags:
        - Roles
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - scope
                - permissions
              properties:
                name:
                  type: string
                description:
                  type: string
                scope:
                  type: string
                  enum:
                    - global
                    - organization
                    - site
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/roles/{id}:
    get:
      operationId: getRole
      summary: Get a role by ID
      tags:
        - Roles
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRole
      summary: Update a role
      tags:
        - Roles
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRole
      summary: Delete a role
      tags:
        - Roles
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      responses:
        204:
          description: Deleted
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/devices:
    get:
      operationId: listDevices
      summary: Paginated list of devices
      tags:
        - Devices
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/limitParam"
        - 
          $ref: "#/components/parameters/offsetParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - devices
                  - count
                  - total
                  - limit
                  - offset
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
                  count:
                    type: integer
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createDevice
      summary: Register a device
      tags:
        - Devices
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - organizationId
                - kind
              properties:
                organizationId:
                  type: string
                kind:
                  type: string
                  enum:
                    - tablet
                    - phone
                    - kiosk
                    - sensor
                    - laptop
                status:
                  type: string
                  enum:
                    - provisioned
                    - active
                    - lost
                    - retired
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Role:
      type: object
      required:
        - id
        - name
        - scope
        - permissions
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        scope:
          type: string
          enum:
            - global
            - organization
            - site
        permissions:
          type: array
          items:
            type: string
    Device:
      type: object
      required:
        - id
        - organizationId
        - kind
        - status
      properties:
        id:
          type: string
        organizationId:
          type: string
        kind:
          type: string
          enum:
            - tablet
            - phone
            - kiosk
            - sensor
            - laptop
        status:
          type: string
          enum:
            - provisioned
            - active
            - lost
            - retired
...
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 200
      description: Number of items per page
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key in the format `keyId:rawSecret`. The server validates via HMAC-SHA256 and constant-time comparison."
      bearerFormat: "keyId:rawSecret"
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by POST /api/v1/auth/token. Include in Authorization header.